Document the signed-ness mismatch between vendor ids
The vendor and product ids are all signed integers except in the
libwacom_match code where they are uint32_t. We can't change that
without API changes (not ABI though), so let's just document it and move on.

+++ test/test-load.c
@@ -49,9 +49,9 @@ static void check_multiple_match(WacomDevice *device)
 		nmatches++;
 		if (libwacom_match_get_bustype(*match) == libwacom_get_bustype(device))
 			found_bus = 1;
-		if (libwacom_match_get_vendor_id(*match) == libwacom_get_vendor_id(device))
+		if ((int)libwacom_match_get_vendor_id(*match) == libwacom_get_vendor_id(device))
 			found_vendor_id = 1;
-		if (libwacom_match_get_product_id(*match) == libwacom_get_product_id(device))
+		if ((int)libwacom_match_get_product_id(*match) == libwacom_get_product_id(device))
 			found_product_id = 1;
 	}
